backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pydantic==2.5.2
pdfplumber==0.10.3
pandas==2.1.3
openpyxl==3.1.2
python-multipart==0.0.6
pdf2image==1.17.0
openai>=1.50.0
httpx>=0.27.0
pytesseract==0.3.13

